+++ webservice/lib_collections.py
@@ -1,13 +1,30 @@
-import json
-import re
 from webservice.lib_misc import parseECLI
-from webservice.lib_async_tools import urlIsPdf
 import logging
+from webservice.sources import GHCC
+from webservice.sources import JUST
+from webservice.sources import RVSCDE
+from webservice.sources import OPENJUSTICE
 logger = logging.getLogger(__name__)
 
 DEFAULT = 'default'
 PDF = 'pdf'
 META = 'meta'
+
+
+def sources():
+    # FIXME: dirty, make cleaner ;)
+    # XXX: Improvement: add params to do checks within this function
+    # Instead of outside
+    yield RVSCDE
+    yield GHCC
+    yield OPENJUSTICE
+    yield JUST
+
+
+for court in sources():
+    court.init()
+    court.init()
+    court.init()
 
 
 def root(config):
@@ -17,15 +34,6 @@ def root(config):
          'rel':''}
         for x in config['countries']
     ]
-
-
-def sources():
-    # FIXME: dirty, make cleaner ;)
-    # XXX: Improvement: add params to do checks within this function
-    # Instead of outside
-    yield RVSCDE
-    yield GHCC
-    yield JUST
 
 
 def getCourt(config, country, court):
@@ -76,7 +84,7 @@ def listDocuments(config, country, court, year):
     docs = set()
     for source in sources():
         if court in source.getCodes(config) and year in source.getYears(config, court):
-            docs = docs.union(source.getDocuments(config, court, year))
+            docs = docs.union(source.getDocuments(config, court, int(year)))
 
     result = [
         {'name': name,
@@ -110,261 +118,14 @@ def getECLICourt(config, ecli):
         logger.info('GHCC ECLI match')
         return GHCC
 
+    if ecli.court in OPENJUSTICE.getCodes(config) and OPENJUSTICE.docMatch(config, ecli.num):
+        logger.info('OPENJUSTICE ECLI match')
+        return OPENJUSTICE
+
     if ecli.court in JUST.getCodes(config):
         return JUST
 
     return JUST
 
 
-def urlGetType(ftype, urls):
-    for url in urls:
-        if url['rel'] == ftype:
-            return url['href']
-    return False
 
-
-class JUST:
-    country = 'BE'
-    data = []
-
-    @staticmethod
-    def init():
-        """
-        FIXME: This is purely experimental. Prepare some persistant storage for this.
-        {"num": 208168, "year": 2010, "language": "french", "type": "arr"}
-        {"num": 72214, "year": 1998, "language": "dutch", "type": "arr"}
-        """
-        data = {}
-
-        with open('./resources/IUBEL.txt', 'r') as f:
-            for line in f:
-                ecli = parseECLI(line, True)
-                if ecli:
-                    y = ecli.year
-                    c = ecli.court
-                    if c in data:
-                        if y in data[c]:
-                            if ecli.num not in data[c][y]:
-                                data[c][y].append(ecli.num)
-                        else:
-                            data[c][y] = [ecli.num]
-                    else:
-                        data[c] = {y: [ecli.num]}
-
-        JUST.data = data
-
-    @staticmethod
-    async def getUrls(config, eclip, forceType=None):
-        # TODO: check if pdf exists
-        # https://iubel.be/IUBELwork/ECLI:BE:AHANT:2004:ARR.20040604.5_NL.pdf
-        urls = [
-            {'rel': 'default', 'href': f"https://iubel.be/IUBELcontent/ViewDecision.php?id={eclip.raw}"},
-            {'rel': 'meta', 'href': f"https://iubel.be/IUBELcontent/ViewDecision.php?id={eclip.raw}"}
-        ]
-
-        test_urls = [
-            f"https://iubel.be/IUBELwork/{eclip.raw}_NL.pdf",
-            f"https://iubel.be/IUBELwork/{eclip.raw}_FR.pdf",
-        ]
-
-        for url in test_urls:
-            if await urlIsPdf(url):
-                urls.append({'rel': 'pdf', 'href': url})
-
-        if not forceType:
-            return urls
-
-        return urlGetType(forceType, urls)
-
-    @staticmethod
-    def getCodes(config):
-        return list(JUST.data.keys())
-
-    @staticmethod
-    def getYears(config, code):
-        return JUST.data[code].keys()
-
-    @staticmethod
-    def checkYear(year, code):
-        return year in JUST.data[code].keys()
-
-    @staticmethod
-    def getDocData(config, eclip):
-        return {
-            'logo': 'https://www.rechtbanken-tribunaux.be/themes/custom/hoverech/logo.svg',
-            'website': 'https://iubel.be/IUBELhome/welkom',
-            'court': eclip.court,
-            'year': eclip.year,
-            'source': 'IUBEL',
-            'decision': eclip.num,
-        }
-
-    @staticmethod
-    def getDocuments(config, code, year):
-        return JUST.data[code][year]
-
-
-class RVSCDE:
-    country = 'BE'
-    code = 'RSCE'
-    data = []
-
-    @staticmethod
-    def docMatch(config, num):
-        mask = re.compile('^\w{3}\.\d{1,6}$')
-        return bool(mask.match(num))
-
-    @staticmethod
-    def getYears(config, _code):
-        return RVSCDE.data.keys()
-
-    @staticmethod
-    async def getUrls(config, eclip, forceType=None):
-        name = eclip.num.split('.')
-        arr_num = name[1]
-        urls = [{
-            'rel': 'default',
-            'href': f"http://www.raadvst-consetat.be/arr.php?nr={arr_num}"
-        }, {
-            'rel': 'pdf',
-            'href': f"http://www.raadvst-consetat.be/arr.php?nr={arr_num}"
-        }]
-
-        if not forceType:
-            return urls
-
-        return urlGetType(forceType, urls)
-
-    @staticmethod
-    def getCodes(config):
-        return [RVSCDE.code]
-
-    @staticmethod
-    def checkYear(year, _code):
-        return year in RVSCDE.data.keys()
-
-    @staticmethod
-    def getDocuments(config, code, year):
-        collection = []
-        for record in RVSCDE.data[year]:
-            name = '{dtype}.{num}'.format(
-                dtype=record['type'].upper(),
-                num=record['num']
-            )
-            collection.append(name)
-
-        return collection
-
-    @staticmethod
-    def getDocData(config, eclip):
-        return {
-            'logo': 'http://www.raadvst-consetat.be/a/s/logo.gif',
-            'website': 'http://www.raadvst-consetat.be/',
-            'court': RVSCDE.code,
-            'source': 'RSCE',
-            'year': eclip.year,
-            'decision': eclip.num,
-        }
-
-    @staticmethod
-    def init():
-        """
-        FIXME: This is purely experimental. Prepare some persistant storage for this.
-        """
-        data = {}
-        with open('./resources/RVSCDE_def.json', 'r') as f:
-            for line in f:
-                rec = json.loads(line)
-                index = str(rec['year'])
-                if index in data:
-                    data[index].append(rec)
-                else:
-                    data[index] = [rec]
-        RVSCDE.data = data
-
-
-class GHCC:
-    country = 'BE'
-    code = 'GHCC'
-    data = []
-
-    @staticmethod
-    def docMatch(config, num):
-        mask = re.compile('^\d{4}\.\d{1,4}(f|n|d)?$')
-        return bool(mask.match(num))
-
-    @staticmethod
-    def getYears(config, _code):
-        return GHCC.data.keys()
-
-    @staticmethod
-    async def getUrls(config, eclip, forceType=None):
-        name = eclip.num.split('.')
-        arr_num = name[1]
-        urls = [{
-            'rel': 'default',
-            'href': f"https://www.const-court.be/public/f/{eclip.year}/{eclip.year}-{arr_num}.pdf"
-        }, {
-            'rel': 'pdf',
-            'href': f"https://www.const-court.be/public/f/{eclip.year}/{eclip.year}-{arr_num}.pdf"
-        }]
-        if not forceType:
-            return urls
-
-        return urlGetType(forceType, urls)
-
-    @staticmethod
-    def getCodes(config):
-        return [GHCC.code]
-
-    @staticmethod
-    def checkYear(year, _code):
-        return year in GHCC.data.keys()
-
-    @staticmethod
-    def getDocuments(config, code, year):
-        collection = []
-        for record in GHCC.data[year]:
-            name = '{year}.{num}{lang}'.format(
-                year=record['year'],
-                num=record['num'],
-                lang='f' if record['language'] == 'french' else 'n'
-            )
-            collection.append(name)
-
-        return collection
-
-    @staticmethod
-    def getDocData(config, eclip):
-        return {
-            'logo': 'https://www.const-court.be/images/titre_index3.gif',
-            'website': 'https://www.const-court.be/',
-            'court': GHCC.code,
-            'source': 'GHCC',
-            'year': eclip.year,
-            'decision': eclip.num,
-        }
-
-    @staticmethod
-    def init():
-        """
-        FIXME: This is purely experimental. Prepare some persistant storage for this.
-        {"num": 208168, "year": 2010, "language": "french", "type": "arr"}
-        {"num": 72214, "year": 1998, "language": "dutch", "type": "arr"}
-        """
-        data = {}
-        with open('./resources/GHCC_def.json', 'r') as f:
-            for line in f:
-                rec = json.loads(line)
-                index = str(rec['year'])
-                if index in data:
-                    data[index].append(rec)
-                else:
-                    data[index] = [rec]
-        GHCC.data = data
-
-
-for court in sources():
-    court.init()
-    court.init()
-    court.init()